latest

+++ common/models/Article.php
@@ -253,7 +253,7 @@ class Article extends \yii\db\ActiveRecord
     public static function getGallery()
     {
         $result = [];
-        $id  = yii::$app->request->get('id');
+        $id  = (int)yii::$app->request->get('id');
         $dir = Yii::getAlias( '@backend' ).'/web/elfinder/global/article_'.$id;
 
         if(file_exists($dir)) {
